Fix scoping bug
We recently ran into a bug with this project in the causes app where
all calls to the Facebook API, which go through method proxy, were
broken for all versions of Internet Explorer. The behavior was
inconsistent in other browsers.

It turns out that this was a scoping problem. The `item` variable
happened to be referencing a global object, causing the unexpected
behavior. The fix for this is to declare the `item` variable in the
local scope.

Change-Id: Ie972fe2e0907182299b087057f406e5b8e718259